requirements_platform.txt
# Platform-specific requirements for RecurrentPPO training

# ============================================================================
# CORE REQUIREMENTS (All platforms)
# ============================================================================

# Essential RL dependencies
stable-baselines3>=2.0.0
sb3_contrib>=2.0.0
gymnasium>=0.28.0

# PyTorch (CPU version for compatibility)
torch>=1.12.0
torchvision>=0.13.0

# Data processing and utilities
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
pyyaml>=6.0

# Logging and visualization
tensorboard>=2.8.0
matplotlib>=3.5.0
tqdm>=4.62.0

# Optional: Weights & Biases logging
# wandb>=0.12.0

# ============================================================================
# LINUX-SPECIFIC REQUIREMENTS
# ============================================================================

# FlyCraft simulation environment (Linux only)
# Install manually on Linux systems:
#   git clone https://github.com/flycraft-team/flycraft
#   cd flycraft && pip install -e .

# ============================================================================
# DEMO REQUIREMENTS (Optional - for Streamlit app)
# ============================================================================

# Streamlit demo dependencies (install separately if needed)
# streamlit>=1.20.0
# plotly>=5.10.0
# imageio>=2.19.0
# imageio-ffmpeg>=0.4.7

# ============================================================================
# DEVELOPMENT REQUIREMENTS (Optional)
# ============================================================================

# Development and testing tools
# pytest>=7.0.0
# black>=22.0.0
# flake8>=4.0.0
# mypy>=0.950

# ============================================================================
# INSTALLATION NOTES
# ============================================================================

# macOS Users:
#   FlyCraft is not supported on macOS. You can install the other dependencies
#   to check syntax and run limited tests, but training requires Linux.
#
# Linux Users:
#   Install all dependencies with:
#     pip install -r requirements_recurrent.txt
#     # Then install FlyCraft manually (see above)
#
# GPU Support:
#   For GPU training, install PyTorch with CUDA:
#     pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
